// src/app/core/models/confirmsweet.interface.ts
export interface confirmAlertsweet {
    title: string
    text: string
    icon: 'success' | 'error' | 'warning' | 'info' | 'question';
    showCancelButton: boolean
    confirmButtonColor: string
    cancelButtonColor: string
    confirmButtonText: string
}

export interface confirmsuccessweet {
    title: string
    text: string
    icon: 'success' | 'error' | 'warning' | 'info' | 'question';
}